requirements.txt
chromadb==0.6.3
dependency_injector==4.46.0
fastapi==0.115.12
google_api_python_client==2.164.0
groq==0.20.0
langchain==0.3.21
langchain_community==0.3.20
langchain_groq==0.3.1
langchain_openai==0.3.11
motor==3.7.0
numpy==2.2.4
openai==1.69.0
protobuf==6.30.2
pydantic==2.11.1
pymongo==4.11.3
python-dotenv==1.1.0
sentence_transformers==4.0.1
SQLAlchemy==2.0.40
tiktoken==0.9.0
